pyproject.toml
[project]
name = "demo"
version = "0.1.0"
description = "Wall Construction API - Django demo for irisai integration"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "AMI Team" }
]
dependencies = [
  "loguru==0.7.3",
  "Django==5.2.7",
  "djangorestframework==3.16.0",
  "django-filter>=24.3",
]

[project.optional-dependencies]
dev = [
  "mypy==1.18.2",
  "ruff==0.13.2",
  "django-stubs==5.2.6",
  "djangorestframework-stubs==3.16.4",
]
test = [
  "pytest==8.4.2",
  "pytest-django==4.9.0",
  "pytest-xdist==3.6.1",
  "factory-boy==3.3.3",
  "Faker==33.3.0",
  "pytest-cov==6.0.0",
  "coverage[toml]==7.6.0",
  "pytest-timeout==2.4.0",
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = []

[tool.setuptools.packages.find]
where = ["."]
include = ["apps*", "config*", "tests*"]
exclude = ["*.tests", "*.tests.*"]

[tool.mypy]
python_version = "3.12"
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true

[tool.ruff]
target-version = "py312"

[tool.coverage.run]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "manage.py",
    "module_setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "if TYPE_CHECKING:",
]
